+++ tests/testthat/test-workbook.R
@@ -2,8 +2,8 @@ test_that("workbook object is created", {
 
   x <- suppressWarnings(generate_workbook(as_aftable(demo_df)))
 
-  expect_s4_class(x, class = "Workbook")
-  expect_identical(class(x)[1], "Workbook")
+  expect_s3_class(x, class = c("wbWorkbook","R6"))
+  expect_identical(class(x)[1], "wbWorkbook")
 
 })
 
@@ -20,7 +20,7 @@ test_that("aftable is passed", {
 
 test_that(".stop_bad_input works as intended", {
 
-  wb <- openxlsx::createWorkbook()
+  wb <- openxlsx2::wb_workbook()
   aftable <- as_aftable(demo_df)
 
   expect_error(.stop_bad_input("x", aftable, "cover"))
@@ -32,6 +32,6 @@ test_that("hyperlinks are generated on the cover page", {
 
   # demo dataset has two hyperlinks on the cover
   y <- suppressWarnings(generate_workbook(as_aftable(demo_df)))
-  expect_length(y$worksheets[[1]]$hyperlinks, 2)
+  expect_equal(sum(grepl("HYPERLINK",y$worksheets[[1]]$sheet_data$cc$f)),2)
 
 })